introduced useWalletConnection hook
